vehicles/carla09/requirements.txt
-r https://raw.githubusercontent.com/carla-simulator/carla/0.9.6/PythonAPI/carla/requirements.txt
# carla >=0.9.5, <1.0
pyzmq >=18.1, <19.0
# numpy >=1.16, <2.0
# pyarrow >=0.15.1, <1.0
# v4l2 >=0.2, <1.0